requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
pillow==11.0.0
sqlparse==0.5.3
pandas>=1.3.0
rapidfuzz>=3.1.1
openpyxl>=3.1.2
